Refactor handling of firewall data requests
This change moves the handling of firewall data requests to its own middleware, GetFirewallDataMiddleware. Previously, this was handled in the ConfigurationService file, which was causing clutter and it was hard to maintain and scale. By doing this, the routing logic and data retrieval logic are more separated, which makes the code much cleaner. Furthermore, the version number of the project in nFirewall.csproj was updated to reflect these changes in the software.