Add new feature to manage whitelist and blacklist IPs
This commit adds two new classes for managing whitelist and blacklist IPs, as described in the user story
APPS-123. These classes (WhiteListIpRange and BlacklistIpRange) extend a new abstract class, AddressRangeList,
which contains common functionality for IP range lists. This will allow users to include or exclude IPs, either
as single addresses or ranges, in their firewall settings. This feature was added to provide more flexibility to
user when configuring security settings. As part of this, the CalculateTotalIpRequest data processor was also
updated to ignore any requests originating from IPs in the whitelist.