Refactor data processors to enhance data tracking
The data processor classes `CalculateTotalUrlRequest`, `CalculateTotalIpRequest`, and `CalculateSlowestUrlRequest` have been refactored to store additional information about requests, including the start and end times. This information is stored in new `CalculateTotalUrlData`, `CalculateTotalIpData` and `CalculateSlowestUrlData` classes, and it will provide more insight into how request frequency changes over time.

In addition, a new class `CalculateRequestPerMinutes` was added to calculate the count of requests made per minute. A safety check was also added in `QueueManager` to return a default value when the request queue is empty. Lastly, an unused import was removed from `BlockIpModule.cs`.

The purpose of these changes is to enhance the ability to track and analyse request data for the nFirewall's services. These changes should improve the statistical analysis capability of the application.