Update calculation method and enhance middleware functionality
Updated the way IP request totals are calculated for improved accuracy, and also refactored the GetFirewallDataMiddleware to enhance its functionality. In the DataProcessors/CalculateTotalIpRequest.cs, the IP request totals are now calculated based on the total value instead of the individual value. In the Middlewares/GetFirewallDataMiddleware.cs, a new method 'SendCommands' is created to send responses, and also the typo in 'SendReport' method is corrected. Meanwhile, the version number in the nFirewall.csproj has been upgraded to reflect these changes and improvements.